Consolidate error packages
