--- pkg/controller/error/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["requeue_error.go"],
-    importpath = "sigs.k8s.io/cluster-api/pkg/controller/error",
-    visibility = ["//visibility:public"],
-    deps = ["//vendor/github.com/pkg/errors:go_default_library"],
-)
