+++ pkg/controller/machine/BUILD.bazel
@@ -11,9 +11,9 @@ go_library(
     deps = [
         "//pkg/apis/cluster/common:go_default_library",
         "//pkg/apis/cluster/v1alpha2:go_default_library",
-        "//pkg/controller/error:go_default_library",
         "//pkg/controller/external:go_default_library",
         "//pkg/controller/remote:go_default_library",
+        "//pkg/errors:go_default_library",
         "//pkg/util:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
@@ -44,11 +44,10 @@ go_test(
     deps = [
         "//pkg/apis:go_default_library",
         "//pkg/apis/cluster/v1alpha2:go_default_library",
-        "//pkg/controller/error:go_default_library",
+        "//pkg/errors:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/golang.org/x/net/context:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
