+++ pkg/controller/machine/machine_controller_phases_test.go
@@ -28,7 +28,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha2"
-	capierrors "sigs.k8s.io/cluster-api/pkg/controller/error"
+	capierrors "sigs.k8s.io/cluster-api/pkg/errors"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
