+++ pkg/errors/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "clusters.go",
+        "controllers.go",
         "deployer.go",
         "machines.go",
     ],
